@@ -37,28 +37,28 @@
         "@testing-library/jest-dom": "^6.6.2",
         "@testing-library/react": "^16.2.0",
         "@types/jest": "^29.5.14",
-        "@types/node": "^22.10.7",
+        "@types/node": "^22.13.5",
         "@types/react": "18.3.12",
-        "@typescript-eslint/eslint-plugin": "^8.20.0",
-        "@typescript-eslint/parser": "^8.20.0",
+        "@typescript-eslint/eslint-plugin": "^8.24.0",
+        "@typescript-eslint/parser": "^8.24.0",
         "aws-sdk-client-mock": "^4.1.0",
-        "eslint": "^9.19.0",
-        "eslint-config-prettier": "^9.1.0",
+        "eslint": "^9.21.0",
+        "eslint-config-prettier": "^10.0.1",
         "eslint-plugin-import-newlines": "^1.3.4",
         "eslint-plugin-prettier": "^5.2.3",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
         "jest-junit": "^16.0.0",
         "license-checker": "^25.0.1",
-        "semantic-release": "^24.2.1",
-        "ts-jest": "^29.2.5",
+        "semantic-release": "^24.2.3",
+        "ts-jest": "^29.2.6",
         "ts-node": "^10.9.2",
         "typescript": "^5.7.3",
-        "typescript-eslint": "^8.22.0"
+        "typescript-eslint": "^8.25.0"
     },
     "dependencies": {
         "conventional-changelog-eslint": "^6.0.0",
-        "esbuild": "^0.24.2",
+        "esbuild": "^0.25.0",
         "next": "^15.1.4",
         "nhsuk-react-components": "^5.0.0",
         "react": "^18.3.1",
