@@ -1,9 +1,15 @@
-import "@testing-library/jest-dom"
-import {render, screen, waitFor} from "@testing-library/react"
-import {useRouter} from 'next/navigation'
-import React from "react"
+import "@testing-library/jest-dom";
+import { render, screen, waitFor } from "@testing-library/react";
+import { useNavigate } from "react-router-dom";
+import React from "react";
 
-import {AuthContext} from "@/context/AuthProvider"
+import { AuthContext } from "@/context/AuthProvider";
+
+import axios from "@/helpers/axios";
+jest.mock("@/helpers/axios");
+
+// Tell TypeScript that axios is a mocked version.
+const mockedAxios = axios as jest.Mocked<typeof axios>;
 
 // Mock the card strings, so we have known text for the tests
 jest.mock("@/constants/ui-strings/CardStrings", () => {
@@ -31,61 +37,67 @@ jest.mock("@/constants/ui-strings/CardStrings", () => {
     noAddress: "No address",
     errorDuringRoleSelection: "Error during role selection",
     loadingMessage: "Loading...",
-  }
-
+  };
 
   const EPS_CARD_STRINGS = {
     noOrgName: "NO ORG NAME",
     noODSCode: "No ODS code",
     noRoleName: "No role name",
-    noAddress: "Address not found"
-  }
+    noAddress: "Address not found",
+  };
 
+  return { CHANGE_YOUR_ROLE_PAGE_TEXT, EPS_CARD_STRINGS };
+});
 
-  return {CHANGE_YOUR_ROLE_PAGE_TEXT, EPS_CARD_STRINGS}
-})
-
-// Mock `next/navigation` to prevent errors during component rendering in test
-jest.mock("next/navigation", () => ({
-  usePathname: jest.fn(),
-  useRouter: jest.fn(),
-}))
+// Mock `react-router-dom` to prevent errors during component rendering in test
+jest.mock("react-router-dom", () => ({
+  useLocation: jest.fn(),
+  useNavigate: jest.fn(),
+}));
 
 // Create a global mock for `fetch` to simulate API requests
-const mockFetch = jest.fn()
-global.fetch = mockFetch
+const mockFetch = jest.fn();
+global.fetch = mockFetch;
 
+// Mock the AccessProvider context
 jest.mock("@/context/AccessProvider", () => {
-  const React = require("react")
+  const React = require("react");
 
-  const mockContextValue = {
+  const createMockContext = (overrides = {}) => ({
     noAccess: false,
     singleAccess: false,
-    selectedRole: {
-      role_name: "role_name",
-      role_id: "role_id",
-      org_code: "org_code",
-      org_name: "org_name",
-      site_name: "site_name",
-      site_address: "site_address",
-      uuid: "uuid",
-    },
+    selectedRole: undefined,
+    rolesWithAccess: [],
+    rolesWithoutAccess: [],
+    loading: false,
+    error: null,
     setNoAccess: jest.fn(),
     setSingleAccess: jest.fn(),
     setSelectedRole: jest.fn(),
-  }
+    ...overrides,
+  });
 
-  const MockAccessContext = React.createContext(mockContextValue)
-  const useAccess = () => React.useContext(MockAccessContext)
+  // Create the initial mock context value
+  const mockContextValue = createMockContext();
+
+  const MockAccessContext = React.createContext(mockContextValue);
+  const useAccess = () => React.useContext(MockAccessContext);
+
+  // Add a method to update the mock value for testing
+  const __setMockAccessValue = (overrides = {}) => {
+    Object.assign(mockContextValue, createMockContext(overrides));
+  };
 
   return {
-    __esModule: true,
     AccessContext: MockAccessContext,
     useAccess,
-  }
-})
+    __setMockAccessValue,
+  };
+});
 
-// Default mock values for the `AuthContext` to simulate authentication state
+const { __setMockAccessValue } = require("@/context/AccessProvider");
+
+// Default mock values for the AuthContext to simulate authentication state
 const defaultAuthContext = {
   error: null, // No errors by default
   user: null, // User is initially null (not logged in)
@@ -94,244 +106,245 @@ const defaultAuthContext = {
   accessToken: null, // No access token available
   cognitoSignIn: jest.fn(), // Mock Cognito sign-in function
   cognitoSignOut: jest.fn(), // Mock Cognito sign-out function
-}
+};
 
-import ChangeRolePage from "@/app/changerole/page"
+import ChangeRolePage from "@/pages/ChangeRolePage";
 
 // Utility function to render the component with custom AuthContext overrides
 const renderWithAuth = (authOverrides = {}, accessOverrides = {}) => {
-  const authValue = {...defaultAuthContext, ...authOverrides}
+  const authValue = { ...defaultAuthContext, ...authOverrides };
   return render(
     <AuthContext.Provider value={authValue}>
       <ChangeRolePage />
-    </AuthContext.Provider>
-  )
-}
+    </AuthContext.Provider>,
+  );
+};
 
-import {CHANGE_YOUR_ROLE_PAGE_TEXT} from "@/constants/ui-strings/ChangeRolePageStrings"
-import {EpsSpinnerStrings} from "../constants/ui-strings/EpsSpinnerStrings"
+import { CHANGE_YOUR_ROLE_PAGE_TEXT } from "@/constants/ui-strings/ChangeRolePageStrings";
+import { EpsSpinnerStrings } from "@/constants/ui-strings/EpsSpinnerStrings";
 
 describe("ChangeRolePage", () => {
-  // Clear all mock calls before each test to avoid state leaks
   beforeEach(() => {
-    jest.clearAllMocks()
-  })
+    jest.clearAllMocks();
+    __setMockAccessValue({});
+  });
 
-  it("renders loading state when signed in but fetch hasn't resolved yet", async () => {
-    // Mock fetch to hang indefinitely, simulating a pending request
-    mockFetch.mockImplementation(() => new Promise(() => {}))
-
-    // Render the page with user signed in
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
-
-    // Verify that the loading text appears
-    const loadingText = screen.getByText(EpsSpinnerStrings.loading)
-    expect(loadingText).toBeInTheDocument()
-  })
-
-  it("renders error summary if fetch returns non-200 status", async () => {
-    // Mock fetch to return a 500 status code (server error)
-    mockFetch.mockResolvedValue({status: 500})
+  it("renders loading state when signed in but API call hasn't resolved yet", async () => {
+    __setMockAccessValue({ loading: true });
 
     // Render the page with user signed in
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
+  });
+
+  it("renders error summary if API call returns non-200 status", async () => {
+    __setMockAccessValue({
+      error: "Failed to fetch CPT user info",
+      loading: false,
+    });
+
+    // Render the page with user signed in
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
 
     // Wait for the error message to appear
     await waitFor(() => {
       // Check for error summary heading
       const errorHeading = screen.getByRole("heading", {
         name: CHANGE_YOUR_ROLE_PAGE_TEXT.errorDuringRoleSelection,
-      })
-      expect(errorHeading).toBeInTheDocument()
+      });
+      expect(errorHeading).toBeInTheDocument();
 
       // Check for specific error text
-      const errorItem = screen.getByText("Failed to fetch CPT user info")
-      expect(errorItem).toBeInTheDocument()
-    })
-  })
+      const errorItem = screen.getByText("Failed to fetch CPT user info");
+      expect(errorItem).toBeInTheDocument();
+    });
+  });
 
-  it("renders error summary if fetch returns 200 but no userInfo is present", async () => {
-    // Mock fetch to return 200 OK but with an empty JSON body
-    mockFetch.mockResolvedValue({
-      status: 200,
-      json: async () => ({}), // No `userInfo` key in response
-    })
+  it("renders error summary if API call returns 200 but no userInfo is present", async () => {
+    __setMockAccessValue({
+      loading: false,
+      error: "Failed to fetch CPT user info",
+    });
 
-    // Render the page with user signed in
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
 
-    // Wait for the error message to appear
     await waitFor(() => {
-      // Check for error summary heading
       const errorHeading = screen.getByRole("heading", {
         name: CHANGE_YOUR_ROLE_PAGE_TEXT.errorDuringRoleSelection,
-      })
-      expect(errorHeading).toBeInTheDocument()
+      });
+      expect(errorHeading).toBeInTheDocument();
 
-      // Check for specific error text
-      const errorItem = screen.getByText("Failed to fetch CPT user info")
-      expect(errorItem).toBeInTheDocument()
-    })
-  })
+      const errorItem = screen.getByText("Failed to fetch CPT user info");
+      expect(errorItem).toBeInTheDocument();
+    });
+  });
 
   it("renders the page content when valid userInfo is returned", async () => {
-    // Mock user data to simulate valid API response
-    const mockUserInfo = {
-      roles_with_access: [
+    __setMockAccessValue({
+      loading: false,
+      rolesWithAccess: [
         {
           role_name: "Pharmacist",
           org_name: "Test Pharmacy Org",
           org_code: "ORG123",
-          site_address: "1 Fake Street",
         },
       ],
-      roles_without_access: [
+      rolesWithoutAccess: [
         {
           role_name: "Technician",
-          org_name: "Tech Org",
+          org_name: "Test Pharmacy Org",
           org_code: "ORG456",
-          site_address: "2 Fake Street",
         },
       ],
-    }
-
-    // Mock fetch to return 200 OK with valid userInfo
-    mockFetch.mockResolvedValue({
-      status: 200,
-      json: async () => ({userInfo: mockUserInfo}),
-    })
+    });
 
     // Render the page with user signed in
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
 
     // Wait for the main content to load
     await waitFor(() => {
       // Check for the page heading
-      const heading = screen.getByRole("heading", {level: 1})
-      expect(heading).toHaveTextContent(CHANGE_YOUR_ROLE_PAGE_TEXT.title)
-    })
+      const heading = screen.getByRole("heading", { level: 1 });
+      expect(heading).toHaveTextContent(CHANGE_YOUR_ROLE_PAGE_TEXT.title);
+    });
 
     // Verify the page caption
-    const caption = screen.getByText(CHANGE_YOUR_ROLE_PAGE_TEXT.caption)
-    expect(caption).toBeInTheDocument()
+    expect(
+      screen.getByText(CHANGE_YOUR_ROLE_PAGE_TEXT.caption),
+    ).toBeInTheDocument();
 
-    // Verify the "Roles without access" section (expander)
-    const expanderText = CHANGE_YOUR_ROLE_PAGE_TEXT.roles_without_access_table_title
-    const expander = screen.getByText(expanderText)
-    expect(expander).toBeInTheDocument()
+    // Verify the "Roles without access" section
+    expect(
+      screen.getByText(
+        CHANGE_YOUR_ROLE_PAGE_TEXT.roles_without_access_table_title,
+      ),
+    ).toBeInTheDocument();
 
-    // Check for the table data in "Roles without access"
-    const tableOrg = screen.getByText(/Tech Org \(ODS: ORG456\)/i)
-    expect(tableOrg).toBeInTheDocument()
-    const tableRole = screen.getByText("Technician")
-    expect(tableRole).toBeInTheDocument()
-  })
+    // Check for the table data using test IDs
+    const nameCell = screen.getByTestId("change-role-name-cell");
+    expect(nameCell).toHaveTextContent(`Test Pharmacy Org (ODS: ORG456)`);
+
+    const roleCell = screen.getByTestId("change-role-role-cell");
+    expect(roleCell).toHaveTextContent("Technician");
+  });
 
   it("renders error summary when not signed in", async () => {
-    // Render the page with `isSignedIn` set to false
-    renderWithAuth({isSignedIn: false, error: "Missing access or ID token"})
+    __setMockAccessValue({
+      loading: false,
+      error: "Missing access or ID token",
+    });
 
-    // Wait for the error message to appear
-    await waitFor(() => {
-      // Check for error summary heading
-      const errorHeading = screen.getByRole("heading", {
-        name: CHANGE_YOUR_ROLE_PAGE_TEXT.errorDuringRoleSelection,
-      })
-      expect(errorHeading).toBeInTheDocument()
+    renderWithAuth({ isSignedIn: false });
 
-      const errorItem = screen.getByText("Missing access or ID token")
-      expect(errorItem).toBeInTheDocument()
-    })
-  })
+    const errorHeading = screen.getByRole("heading", {
+      name: CHANGE_YOUR_ROLE_PAGE_TEXT.errorDuringRoleSelection,
+    });
+    expect(errorHeading).toBeInTheDocument();
+    expect(screen.getByText("Missing access or ID token")).toBeInTheDocument();
+  });
 
   it("redirects to searchforaprescription when there is one role with access and no roles without access", async () => {
-    const mockUserInfo = {
-      roles_with_access: [
+    __setMockAccessValue({
+      loading: false,
+      rolesWithAccess: [
         {
           role_name: "Pharmacist",
           org_name: "Test Pharmacy Org",
           org_code: "ORG123",
-          site_address: "1 Fake Street",
         },
       ],
-      roles_without_access: [],
-    }
+      rolesWithoutAccess: [],
+      singleAccess: true,
+    });
 
-    // Mock fetch to return 200 OK with valid userInfo
-    mockFetch.mockResolvedValue({
-      status: 200,
-      json: async () => ({userInfo: mockUserInfo}),
-    })
+    const mockNavigate = jest.fn();
+    (useNavigate as jest.Mock).mockReturnValue(mockNavigate);
 
     // Render the page with user signed in
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
-
-    // Mock useRouter's push function
-    const mockPush = jest.fn();
-    (useRouter as jest.Mock).mockReturnValue({
-      push: mockPush,
-    })
-
-    // Render the page with user signed in
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
 
     // Wait for redirection
     await waitFor(() => {
-      expect(mockPush).toHaveBeenCalledWith("/searchforaprescription")
-    })
-  })
+      expect(mockNavigate).toHaveBeenCalledWith("/searchforaprescription");
+    });
+  });
 
   it("renders loading state when waiting for API response", async () => {
-    mockFetch.mockImplementation(() => new Promise(() => {}))
-    renderWithAuth()
-    expect(screen.getByText("Loading...")).toBeInTheDocument()
-  })
+    __setMockAccessValue({
+      loading: true,
+    });
+    mockedAxios.get.mockImplementation(() => new Promise(() => {}));
+    renderWithAuth();
+    expect(screen.getByText("Loading...")).toBeInTheDocument();
+  });
 
   it("redirects when a single role is available", async () => {
-    (useRouter as jest.Mock).mockReturnValue({
-      push: jest.fn()
-    })
+    const mockNavigate = jest.fn();
+    (useNavigate as jest.Mock).mockReturnValue(mockNavigate);
 
-    const mockUserInfo = {
-      roles_with_access: [{
-        role_name: "Pharmacist",
-        org_name: "Test Pharmacy",
-        org_code: "ORG123",
-        site_address: "123 Test St"
-      }],
-      roles_without_access: []
-    }
+    __setMockAccessValue({
+      loading: false,
+      rolesWithAccess: [
+        {
+          role_name: "Pharmacist",
+          org_name: "Test Pharmacy",
+          org_code: "ORG123",
+        },
+      ],
+      rolesWithoutAccess: [],
+      singleAccess: true,
+    });
 
-    mockFetch.mockResolvedValue({
-      status: 200,
-      json: async () => ({userInfo: mockUserInfo})
-    })
-
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
 
     await waitFor(() => {
-      expect(useRouter().push).toHaveBeenCalledWith("/searchforaprescription")
-    })
-  })
+      expect(mockNavigate).toHaveBeenCalledWith("/searchforaprescription");
+    });
+  });
 
   it("does not fetch user roles if user is not signed in", async () => {
-    const mockFetch = jest.fn()
-    global.fetch = mockFetch
-
-    renderWithAuth({isSignedIn: false}) // Simulating a user who is not signed in
-
-    expect(mockFetch).not.toHaveBeenCalled()
-  })
+    renderWithAuth({ isSignedIn: false });
+    expect(mockedAxios.get).not.toHaveBeenCalled();
+  });
 
   it("displays an error when the API request fails", async () => {
-    mockFetch.mockRejectedValue(new Error("Failed to fetch user roles"))
+    __setMockAccessValue({
+      loading: false,
+      error: "Failed to fetch CPT user info",
+      rolesWithAccess: [],
+      rolesWithoutAccess: [],
+    });
 
-    renderWithAuth({isSignedIn: true, idToken: {toString: jest.fn().mockReturnValue("mock-id-token")}})
+    renderWithAuth({
+      isSignedIn: true,
+      idToken: { toString: jest.fn().mockReturnValue("mock-id-token") },
+    });
 
     await waitFor(() => {
-      const errorSummary = screen.getByRole("heading", {name: "Error during role selection"})
-      expect(errorSummary).toBeInTheDocument()
-      expect(screen.getByText("Failed to fetch CPT user info")).toBeInTheDocument()
-    })
-  })
-})
+      const errorHeading = screen.getByRole("heading", {
+        name: CHANGE_YOUR_ROLE_PAGE_TEXT.errorDuringRoleSelection,
+      });
+      expect(errorHeading).toBeInTheDocument();
+
+      const errorMessage = screen.getByText("Failed to fetch CPT user info");
+      expect(errorMessage).toBeInTheDocument();
+    });
+  });
+});
