@@ -1,23 +1,65 @@
-import "@testing-library/jest-dom"
-import {render, screen, fireEvent} from "@testing-library/react"
-import EpsCard from "@/components/EpsCard"
-import {useRouter} from "next/navigation"
-import {AuthContext} from "@/context/AuthProvider"
-import {useAccess} from "@/context/AccessProvider"
+import "@testing-library/jest-dom";
+import { render, screen, fireEvent } from "@testing-library/react";
+import { BrowserRouter } from "react-router-dom";
+import EpsCard from "@/components/EpsCard";
+import { AuthContext } from "@/context/AuthProvider";
+import { AccessContext } from "@/context/AccessProvider";
+import { JWT } from "aws-amplify/auth";
 
-// Mock `next/navigation`
-jest.mock("next/navigation", () => ({
-  useRouter: jest.fn(),
-}))
+// Mock the auth configuration
+jest.mock("@/context/configureAmplify", () => ({
+  __esModule: true,
+  authConfig: {
+    Auth: {
+      Cognito: {
+        userPoolClientId: "mockUserPoolClientId",
+        userPoolId: "mockUserPoolId",
+        loginWith: {
+          oauth: {
+            domain: "mockHostedLoginDomain",
+            scopes: [
+              "openid",
+              "email",
+              "phone",
+              "profile",
+              "aws.cognito.signin.user.admin",
+            ],
+            redirectSignIn: ["mockRedirectSignIn"],
+            redirectSignOut: ["mockRedirectSignOut"],
+            responseType: "code",
+          },
+          username: true,
+          email: false,
+          phone: false,
+        },
+      },
+    },
+  },
+}));
 
-// Mock `useAccess` hook
-jest.mock("@/context/AccessProvider", () => ({
-  useAccess: jest.fn(),
-}))
+// Mock useNavigate
+const mockNavigate = jest.fn();
+jest.mock("react-router-dom", () => ({
+  ...jest.requireActual("react-router-dom"),
+  useNavigate: () => mockNavigate,
+}));
 
-// Mock global fetch
-const mockFetch = jest.fn()
-global.fetch = mockFetch as jest.Mock
+// Mock EPS_CARD_STRINGS
+jest.mock("@/constants/ui-strings/CardStrings", () => ({
+  EPS_CARD_STRINGS: {
+    noODSCode: "No ODS code",
+    noOrgName: "NO ORG NAME",
+    noRoleName: "No role name",
+    noAddress: "Address not found",
+  },
+}));
+
+jest.mock("@/constants/environment", () => ({
+  API_ENDPOINTS: {
+    TRACKER_USER_INFO: "/mock-endpoint",
+    SELECTED_ROLE: "/mock-endpoint",
+  },
+}));
 
 const mockRole = {
   role_id: "123",
@@ -25,71 +67,135 @@ const mockRole = {
   org_code: "XYZ123",
   role_name: "Pharmacist",
   site_address: "123 Test Street\nTest City",
-}
+};
 
-const mockLink = "/role-detail"
+const mockLink = "/role-detail";
+
+// Default context values
+const mockJWT = {
+  toString: () => "mock-token",
+  payload: {
+    sub: "mock-sub",
+    aud: "mock-audience",
+    iss: "mock-issuer",
+    token_use: "id",
+  },
+} satisfies JWT;
+
+// Update default auth context with proper JWT
+const defaultAuthContext = {
+  error: null,
+  user: null,
+  isSignedIn: true,
+  idToken: "mock-token",
+  accessToken: null,
+  cognitoSignIn: jest.fn(),
+  cognitoSignOut: jest.fn(),
+};
+
+const defaultAccessContext = {
+  setSelectedRole: jest.fn(),
+  noAccess: false,
+  singleAccess: false,
+  selectedRole: null,
+  userDetails: undefined,
+  setUserDetails: jest.fn(),
+  setNoAccess: jest.fn(),
+  setSingleAccess: jest.fn(),
+  clear: jest.fn(),
+};
+
+const renderWithProviders = (
+  props: { role: any; link: string },
+  authOverrides: Partial<React.ContextType<typeof AuthContext>> = {},
+  accessOverrides: Partial<React.ContextType<typeof AccessContext>> = {},
+) => {
+  const authValue = {
+    ...defaultAuthContext,
+    ...authOverrides,
+  } as React.ContextType<typeof AuthContext>;
+  const accessValue = {
+    ...defaultAccessContext,
+    ...accessOverrides,
+  } as React.ContextType<typeof AccessContext>;
+
+  return render(
+    <AuthContext.Provider value={authValue}>
+      <AccessContext.Provider value={accessValue}>
+        <BrowserRouter>
+          <EpsCard {...props} />
+        </BrowserRouter>
+      </AccessContext.Provider>
+    </AuthContext.Provider>,
+  );
+};
 
 describe("EpsCard Component", () => {
-  let mockRouterPush: jest.Mock
-  let mockSetSelectedRole: jest.Mock
-
   beforeEach(() => {
-    jest.clearAllMocks()
-    mockRouterPush = jest.fn();
-    (useRouter as jest.Mock).mockReturnValue({push: mockRouterPush})
-
-    mockSetSelectedRole = jest.fn();
-    (useAccess as jest.Mock).mockReturnValue({setSelectedRole: mockSetSelectedRole})
-  })
+    jest.clearAllMocks();
+    global.fetch = jest.fn().mockImplementation(() =>
+      Promise.resolve({
+        ok: true,
+        json: () => Promise.resolve({ message: "Success" }),
+      }),
+    );
+  });
 
   it("renders role details correctly", () => {
-    render(<EpsCard role={mockRole} link={mockLink} />)
+    renderWithProviders({ role: mockRole, link: mockLink });
 
-    expect(screen.getByText((content) => content.includes("Test Organization"))).toBeInTheDocument()
-    expect(screen.getByText((content) => content.includes("ODS: XYZ123"))).toBeInTheDocument()
-    expect(screen.getByText("Pharmacist")).toBeInTheDocument()
-    expect(screen.getByText("123 Test Street", {exact: false})).toBeInTheDocument()
-    expect(screen.getByText("Test City", {exact: false})).toBeInTheDocument()
-  })
+    expect(
+      screen.getByText((content) => content.includes("Test Organization")),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByText((content) => content.includes("ODS: XYZ123")),
+    ).toBeInTheDocument();
+    expect(screen.getByText("Pharmacist")).toBeInTheDocument();
+    expect(screen.getByText("123 Test Street")).toBeInTheDocument();
+    expect(screen.getByText("Test City")).toBeInTheDocument();
+  });
 
   it("handles missing role data gracefully", () => {
-    render(<EpsCard role={{} as any} link={mockLink} />)
+    renderWithProviders({ role: {}, link: mockLink });
 
-    expect(screen.getByText((content) => content.includes("NO ORG NAME"))).toBeInTheDocument()
-    expect(screen.getByText((content) => content.includes("ODS: No ODS code"))).toBeInTheDocument()
-    expect(screen.getByText("No role name")).toBeInTheDocument()
-    expect(screen.getByText("Address not found", {exact: false})).toBeInTheDocument()
-  })
+    expect(
+      screen.getByText((content) => content.includes("NO ORG NAME")),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByText((content) => content.includes("ODS: No ODS code")),
+    ).toBeInTheDocument();
+    expect(screen.getByText("No role name")).toBeInTheDocument();
+    expect(screen.getByText("Address not found")).toBeInTheDocument();
+  });
 
   it("calls API and navigates on card click", async () => {
-    mockFetch.mockResolvedValue({
-      ok: true,
-      json: async () => ({message: "Success"}),
-    })
+    const mockSetSelectedRole = jest.fn();
+    renderWithProviders(
+      { role: mockRole, link: mockLink },
+      {},
+      { setSelectedRole: mockSetSelectedRole },
+    );
 
-    render(<EpsCard role={mockRole} link={mockLink} />)
+    const cardLink = screen.getByRole("link", { name: /test organization/i });
+    await fireEvent.click(cardLink);
 
-    const cardLink = screen.getByRole("link", {name: /test organization/i})
-    await fireEvent.click(cardLink)
-
-    expect(mockFetch).toHaveBeenCalledWith(
-      "/api/selected-role",
-      expect.objectContaining({
-        method: "PUT",
-        headers: expect.objectContaining({
-          "Authorization": expect.any(String),
-          "Content-Type": "application/json",
-        }),
-        body: JSON.stringify({
-          currently_selected_role: {
-            role_id: "123",
-            org_name: "Test Organization",
-            org_code: "XYZ123",
-            role_name: "Pharmacist",
-          },
-        }),
-      })
-    )
+    // Update to match the mocked endpoint
+    expect(global.fetch).toHaveBeenCalledWith("/mock-endpoint", {
+      method: "PUT",
+      headers: {
+        Authorization: "Bearer mock-token",
+        "Content-Type": "application/json",
+        "NHSD-Session-URID": "555254242106",
+      },
+      body: JSON.stringify({
+        currently_selected_role: {
+          role_id: "123",
+          org_name: "Test Organization",
+          org_code: "XYZ123",
+          role_name: "Pharmacist",
+        },
+      }),
+    });
 
     expect(mockSetSelectedRole).toHaveBeenCalledWith(
       expect.objectContaining({
@@ -97,23 +203,29 @@ describe("EpsCard Component", () => {
         org_name: "Test Organization",
         org_code: "XYZ123",
         role_name: "Pharmacist",
-      })
-    )
+      }),
+    );
 
-    expect(mockRouterPush).toHaveBeenCalledWith(mockLink)
-  })
+    expect(mockNavigate).toHaveBeenCalledWith(mockLink);
+  });
 
   it("shows an alert when API call fails", async () => {
-    mockFetch.mockResolvedValueOnce({ok: false})
-    jest.spyOn(window, "alert").mockImplementation(() => {})
+    global.fetch = jest.fn().mockImplementation(() =>
+      Promise.resolve({
+        ok: false,
+      }),
+    );
+    const mockAlert = jest.spyOn(window, "alert").mockImplementation(() => {});
 
-    render(<EpsCard role={mockRole} link={mockLink} />)
+    renderWithProviders({ role: mockRole, link: mockLink });
 
-    const cardLink = screen.getByRole("link", {name: /test organization/i})
-    await fireEvent.click(cardLink)
+    const cardLink = screen.getByRole("link", { name: /test organization/i });
+    await fireEvent.click(cardLink);
 
-    expect(window.alert).toHaveBeenCalledWith("There was an issue selecting your role. Please try again.")
+    expect(mockAlert).toHaveBeenCalledWith(
+      "There was an issue selecting your role. Please try again.",
+    );
 
-    jest.restoreAllMocks()
-  })
-})
+    mockAlert.mockRestore();
+  });
+});
